scripts/requirements-migrate.txt
# Зависимости для migration utility
click>=8.1
psycopg2-binary>=2.9
pyyaml>=6.0
rich>=13.0
pydantic>=2.9
python-dotenv>=1.0
